feat: change 'prf' accepted inputs
The 'prf' parameter now accepts both a PRF function or a character string representing the name or OID of that function.
The paramerter is checked against a list of known functions: this takes away the possibility to plug user-defined functions into pbkdf2, but avoids the misuse of any function (mean, sum, ...) that will previously be acepted as input but will then result in an error.